further split up AppKitBackend
