+++ Sources/AppKitBackend/AppKitBackend+Images.swift
@@ -1,0 +1,40 @@
+import AppKit
+import SwiftCrossUI
+
+extension AppKitBackend {
+    public func createImageView() -> Widget {
+        let imageView = NSImageView()
+        imageView.imageScaling = .scaleAxesIndependently
+        return imageView
+    }
+
+    public func updateImageView(
+        _ imageView: Widget,
+        rgbaData: [UInt8],
+        width: Int,
+        height: Int,
+        targetWidth: Int,
+        targetHeight: Int,
+        dataHasChanged: Bool,
+        environment: EnvironmentValues
+    ) {
+        guard dataHasChanged else {
+            return
+        }
+
+        let imageView = imageView as! NSImageView
+        var rgbaData = rgbaData
+        let context = CGContext(
+            data: &rgbaData,
+            width: width,
+            height: height,
+            bitsPerComponent: 8,
+            bytesPerRow: 4 * width,
+            space: CGColorSpaceCreateDeviceRGB(),
+            bitmapInfo: CGImageAlphaInfo.premultipliedLast.rawValue
+        )
+        let cgImage = context!.makeImage()!
+
+        imageView.image = NSImage(cgImage: cgImage, size: NSSize(width: width, height: height))
+    }
+}
