+++ Sources/AppKitBackend/AppKitBackend.swift
@@ -12,7 +12,6 @@ extension App {
 public final class AppKitBackend: FullAppBackend {
     public typealias Window = NSCustomWindow
     public typealias Widget = NSView
-    public typealias Alert = NSAlert
 
     public let defaultTableRowContentHeight = 20
     public let defaultTableCellVerticalPadding = 4
@@ -21,12 +20,6 @@ public final class AppKitBackend: FullAppBackend {
     public let requiresImageUpdateOnScaleFactorChange = false
     public let supportsMultipleWindows = true
     public let deviceClass = DeviceClass.desktop
-    public let supportedDatePickerStyles: [DatePickerStyle] = [.automatic, .graphical, .compact]
-    public let supportedPickerStyles: [BackendPickerStyle] = [
-        .menu,
-        .segmented,
-        .radioGroup,
-    ]
     public let canOverrideWindowColorScheme = true
 
     public var scrollBarWidth: Int {
@@ -170,15 +163,6 @@ public final class AppKitBackend: FullAppBackend {
         window.makeKeyAndOrderFront(nil)
     }
 
-    public func setApplicationMenu(
-        _ submenus: [ResolvedMenu.Submenu],
-        environment: EnvironmentValues
-    ) {
-        MenuBar.setUpMenuBar(extraMenus: submenus.map {
-            Self.renderSubmenu($0, environment: environment)
-        })
-    }
-
     public func close(window: Window) {
         window.close()
     }
@@ -196,73 +180,6 @@ public final class AppKitBackend: FullAppBackend {
 
     public func revealFile(_ url: URL) throws {
         NSWorkspace.shared.activateFileViewerSelecting([url])
-    }
-
-    static func renderMenuItem(
-        _ item: ResolvedMenu.Item,
-        environment: EnvironmentValues
-    ) -> NSMenuItem {
-        switch item {
-            case .button(let label, let action):
-                // Custom subclass is used to keep strong reference to action
-                // wrapper.
-                let renderedItem = NSCustomMenuItem(
-                    title: label,
-                    action: nil,
-                    keyEquivalent: ""
-                )
-                if let action, environment.isEnabled {
-                    let wrappedAction = Action(action)
-                    renderedItem.actionWrapper = wrappedAction
-                    renderedItem.action = #selector(wrappedAction.run)
-                    renderedItem.target = wrappedAction
-                }
-                return renderedItem
-            case .toggle(let label, let value, let onChange):
-                // Custom subclass is used to keep strong reference to action
-                // wrapper.
-                let renderedItem = NSCustomMenuItem(
-                    title: label,
-                    action: nil,
-                    keyEquivalent: ""
-                )
-                renderedItem.isOn = value
-
-                if environment.isEnabled {
-                    let wrappedAction = Action {
-                        onChange(!renderedItem.isOn)
-                    }
-                    renderedItem.actionWrapper = wrappedAction
-                    renderedItem.action = #selector(wrappedAction.run)
-                    renderedItem.target = wrappedAction
-                }
-
-                return renderedItem
-            case .separator:
-                return NSCustomMenuItem.separator()
-            case .submenu(let submenu):
-                return renderSubmenu(submenu, environment: environment)
-            case .modifiedEnvironment(let item, let modification):
-                return renderMenuItem(
-                    item,
-                    environment: modification(environment)
-                )
-        }
-    }
-
-    static func renderSubmenu(
-        _ submenu: ResolvedMenu.Submenu,
-        environment: EnvironmentValues
-    ) -> NSMenuItem {
-        let renderedMenu = NSMenu()
-        renderedMenu.items = submenu.content.items.map {
-            Self.renderMenuItem($0, environment: environment)
-        }
-
-        let menuItem = NSMenuItem()
-        menuItem.title = submenu.label
-        menuItem.submenu = renderedMenu
-        return menuItem
     }
 
     public func runInMainThread(action: @escaping @MainActor () -> Void) {
@@ -550,898 +467,6 @@ public final class AppKitBackend: FullAppBackend {
     public func updateTooltipContainer(_ widget: NSView, tooltip: String) {
         widget.toolTip = tooltip
     }
-
-    public func size(
-        of text: String,
-        whenDisplayedIn widget: Widget,
-        proposedWidth: Int?,
-        proposedHeight: Int?,
-        environment: EnvironmentValues
-    ) -> SIMD2<Int> {
-        let proposedSize = NSSize(
-            width: proposedWidth.map(Double.init) ?? .greatestFiniteMagnitude,
-            height: proposedHeight.map(Double.init) ?? .greatestFiniteMagnitude
-        )
-        let rect = NSString(string: text).boundingRect(
-            with: proposedSize,
-            options: [.usesLineFragmentOrigin, .truncatesLastVisibleLine],
-            attributes: Self.attributes(forTextIn: environment)
-        )
-
-        var height = rect.size.height
-
-        if let lineLimitSettings = environment.lineLimitSettings {
-            let limitedHeight =
-                Double(max(lineLimitSettings.limit, 1)) * environment.resolvedFont.lineHeight
-
-            if limitedHeight < height || lineLimitSettings.reservesSpace {
-                height = limitedHeight
-            }
-        }
-
-        return SIMD2(
-            Int(rect.size.width.rounded(.awayFromZero)),
-            Int(height.rounded(.awayFromZero))
-        )
-    }
-
-    public func createTextView() -> Widget {
-        let field = NSTextField(wrappingLabelWithString: "")
-        // Somewhat unintuitively, this changes the behaviour of the text field even
-        // though it's not editable. It prevents the text from resetting to default
-        // styles when clicked (yeah that happens...)
-        field.allowsEditingTextAttributes = true
-        field.isSelectable = false
-        field.cell?.truncatesLastVisibleLine = true
-        return field
-    }
-
-    public func updateTextView(
-        _ textView: Widget,
-        content: String,
-        environment: EnvironmentValues
-    ) {
-        let field = textView as! NSTextField
-        field.attributedStringValue = Self.attributedString(for: content, in: environment)
-        if field.isSelectable && !environment.isTextSelectionEnabled {
-            field.abortEditing()
-        }
-        field.isSelectable = environment.isTextSelectionEnabled
-    }
-
-    public func createButton() -> Widget {
-        return NSButton(title: "", target: nil, action: nil)
-    }
-
-    public func updateButton(
-        _ button: Widget,
-        label: String,
-        environment: EnvironmentValues,
-        action: @escaping () -> Void
-    ) {
-        let button = button as! NSButton
-        button.attributedTitle = Self.attributedString(
-            for: label,
-            in: environment.with(\.multilineTextAlignment, .center)
-        )
-        button.bezelStyle = .regularSquare
-        button.appearance = environment.colorScheme.nsAppearance
-        button.isEnabled = environment.isEnabled
-        button.onAction = { _ in
-            action()
-        }
-    }
-
-    public func createSwitch() -> Widget {
-        return NSSwitch()
-    }
-
-    public func updateSwitch(
-        _ toggleSwitch: Widget,
-        environment: EnvironmentValues,
-        onChange: @escaping (Bool) -> Void
-    ) {
-        let toggleSwitch = toggleSwitch as! NSSwitch
-        toggleSwitch.isEnabled = environment.isEnabled
-        toggleSwitch.onAction = { toggleSwitch in
-            let toggleSwitch = toggleSwitch as! NSSwitch
-            onChange(toggleSwitch.state == .on)
-        }
-    }
-
-    public func setState(ofSwitch toggleSwitch: Widget, to state: Bool) {
-        let toggleSwitch = toggleSwitch as! NSSwitch
-        toggleSwitch.state = state ? .on : .off
-    }
-
-    public func createToggle() -> Widget {
-        let toggle = NSButton()
-        toggle.setButtonType(.pushOnPushOff)
-        return toggle
-    }
-
-    public func updateToggle(
-        _ toggle: Widget,
-        label: String,
-        environment: EnvironmentValues,
-        onChange: @escaping (Bool) -> Void
-    ) {
-        let toggle = toggle as! NSButton
-        toggle.attributedTitle = Self.attributedString(
-            for: label,
-            in: environment.with(\.multilineTextAlignment, .center)
-        )
-        toggle.isEnabled = environment.isEnabled
-        toggle.onAction = { toggle in
-            let toggle = toggle as! NSButton
-            onChange(toggle.state == .on)
-        }
-    }
-
-    public func setState(ofToggle toggle: Widget, to state: Bool) {
-        let toggle = toggle as! NSButton
-        toggle.state = state ? .on : .off
-    }
-
-    public func createCheckbox() -> Widget {
-        NSButton(checkboxWithTitle: "", target: nil, action: nil)
-    }
-
-    public func updateCheckbox(
-        _ checkbox: Widget,
-        environment: EnvironmentValues,
-        onChange: @escaping (Bool) -> Void
-    ) {
-        let checkbox = checkbox as! NSButton
-        checkbox.isEnabled = environment.isEnabled
-        checkbox.onAction = { toggle in
-            let checkbox = toggle as! NSButton
-            onChange(checkbox.state == .on)
-        }
-    }
-
-    public func setState(ofCheckbox checkbox: Widget, to state: Bool) {
-        let toggle = checkbox as! NSButton
-        toggle.state = state ? .on : .off
-    }
-
-    public func createSlider() -> Widget {
-        return NSSlider()
-    }
-
-    public func updateSlider(
-        _ slider: Widget,
-        minimum: Double,
-        maximum: Double,
-        decimalPlaces: Int,
-        environment: EnvironmentValues,
-        onChange: @escaping (Double) -> Void
-    ) {
-        // TODO: Implement decimalPlaces
-        let slider = slider as! NSSlider
-        slider.minValue = minimum
-        slider.maxValue = maximum
-        slider.onAction = { slider in
-            let slider = slider as! NSSlider
-            onChange(slider.doubleValue)
-        }
-        slider.isEnabled = environment.isEnabled
-    }
-
-    public func setValue(ofSlider slider: Widget, to value: Double) {
-        let slider = slider as! NSSlider
-        slider.doubleValue = value
-    }
-
-    public func createPicker(style: BackendPickerStyle) -> Widget {
-        switch style {
-            case .menu:
-                return NSPopUpButton()
-            case .segmented:
-                return NSSegmentedControl()
-            case .radioGroup:
-                return RadioGroup()
-            default:
-                let message = "unsupported picker style \(style)"
-                logger.critical("\(message)")
-                fatalError(message)
-        }
-    }
-
-    public func updatePicker(
-        _ picker: Widget,
-        options: [String],
-        environment: EnvironmentValues,
-        onChange: @escaping (Int?) -> Void
-    ) {
-        if let picker = picker as? NSPopUpButton {
-            picker.isEnabled = environment.isEnabled
-
-            let menu = picker.menu!
-
-            for (item, option) in zip(menu.items, options) {
-                item.attributedTitle = Self.attributedString(for: option, in: environment)
-            }
-
-            if menu.numberOfItems < options.count {
-                for i in menu.numberOfItems..<options.count {
-                    let item = NSMenuItem()
-                    item.attributedTitle = Self.attributedString(for: options[i], in: environment)
-                    menu.addItem(item)
-                }
-            } else {
-                for i in (options.count..<menu.numberOfItems).reversed() {
-                    menu.removeItem(at: i)
-                }
-            }
-
-            picker.onAction = { picker in
-                let picker = picker as! NSPopUpButton
-                onChange(picker.indexOfSelectedItem)
-            }
-            picker.bezelStyle = .regularSquare
-        } else if let picker = picker as? NSSegmentedControl {
-            picker.isEnabled = environment.isEnabled
-            picker.segmentCount = options.count
-            for (i, option) in options.enumerated() {
-                picker.setLabel(option, forSegment: i)
-            }
-            picker.onAction = { picker in
-                let picker = picker as! NSSegmentedControl
-                let selectedIndex = picker.selectedSegment
-                onChange(selectedIndex == -1 ? nil : selectedIndex)
-            }
-        } else if let picker = picker as? RadioGroup {
-            picker.update(options: options, environment: environment)
-            picker.onChange = onChange
-        }
-    }
-
-    public func setSelectedOption(ofPicker picker: Widget, to selectedOption: Int?) {
-        if let picker = picker as? NSPopUpButton {
-            if let index = selectedOption {
-                picker.selectItem(at: index)
-            } else {
-                picker.select(nil)
-            }
-        } else if let picker = picker as? NSSegmentedControl {
-            picker.selectedSegment = selectedOption ?? -1
-        } else if let picker = picker as? RadioGroup {
-            picker.setSelectedIndex(to: selectedOption)
-        }
-    }
-
-    public func createScrollContainer(for child: Widget) -> Widget {
-        let scrollView = NSScrollView()
-
-        let clipView = scrollView.contentView
-        let documentView = NSStackView()
-        documentView.orientation = .vertical
-        documentView.alignment = .leading
-        documentView.translatesAutoresizingMaskIntoConstraints = false
-        documentView.addView(child, in: .top)
-        scrollView.documentView = documentView
-
-        scrollView.drawsBackground = false
-
-        documentView.topAnchor.constraint(equalTo: clipView.topAnchor).isActive = true
-        documentView.leftAnchor.constraint(equalTo: clipView.leftAnchor).isActive = true
-        documentView.heightAnchor.constraint(greaterThanOrEqualTo: clipView.heightAnchor)
-            .isActive = true
-        documentView.widthAnchor.constraint(greaterThanOrEqualTo: clipView.widthAnchor)
-            .isActive = true
-
-        return scrollView
-    }
-
-    public func updateScrollContainer(
-        _ scrollView: Widget,
-        environment: EnvironmentValues,
-        bounceHorizontally: Bool,
-        bounceVertically: Bool,
-        hasHorizontalScrollBar: Bool,
-        hasVerticalScrollBar: Bool
-    ) {
-        let scrollView = scrollView as! NSScrollView
-        scrollView.hasVerticalScroller = hasVerticalScrollBar
-        scrollView.hasHorizontalScroller = hasHorizontalScrollBar
-        scrollView.verticalScrollElasticity = bounceVertically ? .allowed : .none
-        scrollView.horizontalScrollElasticity = bounceHorizontally ? .allowed : .none
-    }
-
-    public func createSelectableListView() -> Widget {
-        let scrollView = NSDisabledScrollView()
-        scrollView.hasHorizontalScroller = false
-        scrollView.hasVerticalScroller = false
-
-        let listView = NSCustomTableView()
-        listView.delegate = listView.customDelegate
-        listView.dataSource = listView.customDelegate
-        listView.allowsColumnSelection = false
-        let column = NSTableColumn(identifier: NSUserInterfaceItemIdentifier("list-column"))
-        listView.customDelegate.columnCount = 1
-        listView.customDelegate.columnIndices = [
-            ObjectIdentifier(column): 0
-        ]
-        listView.customDelegate.allowSelections = true
-        listView.backgroundColor = .clear
-        listView.headerView = nil
-        listView.addTableColumn(column)
-        if #available(macOS 11.0, *) {
-            listView.style = .plain
-        }
-
-        scrollView.documentView = listView
-        listView.enclosingScrollView?.drawsBackground = false
-        return scrollView
-    }
-
-    public func updateSelectableListView(
-        _ selectableListView: Widget,
-        environment: EnvironmentValues
-    ) {
-        let scrollView = selectableListView as! NSDisabledScrollView
-        let listView = scrollView.documentView! as! NSCustomTableView
-        listView.isEnabled = environment.isEnabled
-    }
-
-    public func baseItemPadding(
-        ofSelectableListView listView: Widget
-    ) -> SwiftCrossUI.EdgeInsets {
-        // TODO: Figure out if there's a way to compute this more directly. At
-        //   the moment these are just figures from empirical observations.
-        SwiftCrossUI.EdgeInsets(top: 0, bottom: 0, leading: 8, trailing: 8)
-    }
-
-    public func minimumRowSize(ofSelectableListView listView: Widget) -> SIMD2<Int> {
-        .zero
-    }
-
-    public func setItems(
-        ofSelectableListView listView: Widget,
-        to items: [Widget],
-        withRowHeights rowHeights: [Int]
-    ) {
-        let listView = (listView as! NSScrollView).documentView! as! NSCustomTableView
-        listView.customDelegate.rowCount = items.count
-        listView.customDelegate.widgets = items
-        listView.customDelegate.rowHeights = rowHeights
-        listView.reloadData()
-    }
-
-    public func setSelectionHandler(
-        forSelectableListView listView: Widget,
-        to action: @escaping (_ selectedIndex: Int) -> Void
-    ) {
-        let listView = (listView as! NSScrollView).documentView! as! NSCustomTableView
-        listView.customDelegate.selectionHandler = action
-    }
-
-    public func setSelectedItem(ofSelectableListView listView: Widget, toItemAt index: Int?) {
-        let listView = (listView as! NSScrollView).documentView! as! NSCustomTableView
-        listView.selectRowIndexes(IndexSet([index].compactMap { $0 }), byExtendingSelection: false)
-    }
-
-    public func createSplitView(leadingChild: Widget, trailingChild: Widget) -> Widget {
-        let splitView = NSCustomSplitView()
-        let leadingChildWithEffect = NSVisualEffectView()
-        leadingChildWithEffect.blendingMode = .behindWindow
-        leadingChildWithEffect.material = .sidebar
-        leadingChildWithEffect.addSubview(leadingChild)
-        leadingChild.widthAnchor.constraint(equalTo: leadingChildWithEffect.widthAnchor)
-            .isActive = true
-        leadingChild.heightAnchor.constraint(equalTo: leadingChildWithEffect.heightAnchor)
-            .isActive = true
-        leadingChild.topAnchor.constraint(equalTo: leadingChildWithEffect.topAnchor)
-            .isActive = true
-        leadingChild.leadingAnchor.constraint(equalTo: leadingChildWithEffect.leadingAnchor)
-            .isActive = true
-        leadingChild.translatesAutoresizingMaskIntoConstraints = false
-        leadingChildWithEffect.translatesAutoresizingMaskIntoConstraints = false
-
-        splitView.addArrangedSubview(leadingChildWithEffect)
-        splitView.addArrangedSubview(trailingChild)
-        splitView.isVertical = true
-        splitView.dividerStyle = .thin
-        let defaultLeadingWidth = 200
-        splitView.setPosition(CGFloat(defaultLeadingWidth), ofDividerAt: 0)
-        splitView.adjustSubviews()
-
-        let delegate = NSSplitViewResizingDelegate()
-        delegate.leadingWidth = defaultLeadingWidth
-        splitView.delegate = delegate
-        splitView.resizingDelegate = delegate
-        return splitView
-    }
-
-    public func setResizeHandler(
-        ofSplitView splitView: Widget,
-        to action: @escaping () -> Void
-    ) {
-        let splitView = splitView as! NSCustomSplitView
-        splitView.resizingDelegate?.setResizeHandler {
-            action()
-        }
-    }
-
-    public func sidebarWidth(ofSplitView splitView: Widget) -> Int {
-        let splitView = splitView as! NSCustomSplitView
-        return splitView.resizingDelegate!.leadingWidth
-    }
-
-    public func setSidebarWidthBounds(
-        ofSplitView splitView: Widget,
-        minimum minimumWidth: Int,
-        maximum maximumWidth: Int
-    ) {
-        let splitView = splitView as! NSCustomSplitView
-        splitView.resizingDelegate!.minimumLeadingWidth = minimumWidth
-        splitView.resizingDelegate!.maximumLeadingWidth = maximumWidth
-    }
-
-    public func createImageView() -> Widget {
-        let imageView = NSImageView()
-        imageView.imageScaling = .scaleAxesIndependently
-        return imageView
-    }
-
-    public func updateImageView(
-        _ imageView: Widget,
-        rgbaData: [UInt8],
-        width: Int,
-        height: Int,
-        targetWidth: Int,
-        targetHeight: Int,
-        dataHasChanged: Bool,
-        environment: EnvironmentValues
-    ) {
-        guard dataHasChanged else {
-            return
-        }
-
-        let imageView = imageView as! NSImageView
-        var rgbaData = rgbaData
-        let context = CGContext(
-            data: &rgbaData,
-            width: width,
-            height: height,
-            bitsPerComponent: 8,
-            bytesPerRow: 4 * width,
-            space: CGColorSpaceCreateDeviceRGB(),
-            bitmapInfo: CGImageAlphaInfo.premultipliedLast.rawValue
-        )
-        let cgImage = context!.makeImage()!
-
-        imageView.image = NSImage(cgImage: cgImage, size: NSSize(width: width, height: height))
-    }
-
-    public func createTable() -> Widget {
-        let scrollView = NSScrollView()
-        let table = NSCustomTableView()
-        table.delegate = table.customDelegate
-        table.dataSource = table.customDelegate
-        table.usesAlternatingRowBackgroundColors = true
-        table.rowHeight = CGFloat(
-            defaultTableRowContentHeight + 2 * defaultTableCellVerticalPadding
-        )
-        table.columnAutoresizingStyle = .lastColumnOnlyAutoresizingStyle
-        table.allowsColumnSelection = false
-        scrollView.documentView = table
-        return scrollView
-    }
-
-    public func setRowCount(ofTable table: Widget, to rowCount: Int) {
-        let table = (table as! NSScrollView).documentView as! NSCustomTableView
-        table.customDelegate.rowCount = rowCount
-    }
-
-    public func setColumnLabels(
-        ofTable table: Widget,
-        to labels: [String],
-        environment: EnvironmentValues
-    ) {
-        let table = (table as! NSScrollView).documentView as! NSCustomTableView
-        var columnIndices: [ObjectIdentifier: Int] = [:]
-        let columns = labels.enumerated().map { (i, label) in
-            let column = NSTableColumn(
-                identifier: NSUserInterfaceItemIdentifier("Column \(i)")
-            )
-            column.headerCell = NSTableHeaderCell()
-            column.headerCell.attributedStringValue = Self.attributedString(
-                for: label,
-                in: environment
-            )
-            columnIndices[ObjectIdentifier(column)] = i
-            return column
-        }
-        table.customDelegate.columnIndices = columnIndices
-        for column in table.tableColumns {
-            table.removeTableColumn(column)
-        }
-        table.customDelegate.columnCount = labels.count
-        for column in columns {
-            table.addTableColumn(column)
-        }
-    }
-
-    public func setCells(
-        ofTable table: Widget,
-        to cells: [Widget],
-        withRowHeights rowHeights: [Int]
-    ) {
-        let table = (table as! NSScrollView).documentView as! NSCustomTableView
-        table.customDelegate.widgets = cells
-        table.customDelegate.rowHeights = rowHeights
-        table.reloadData()
-    }
-
-    internal static func attributedString(
-        for text: String,
-        in environment: EnvironmentValues
-    ) -> NSAttributedString {
-        NSAttributedString(
-            string: text,
-            attributes: attributes(forTextIn: environment)
-        )
-    }
-
-    private static func attributes(
-        forTextIn environment: EnvironmentValues
-    ) -> [NSAttributedString.Key: Any] {
-        let paragraphStyle = NSMutableParagraphStyle()
-        paragraphStyle.alignment =
-            switch environment.multilineTextAlignment {
-                case .leading:
-                    .left
-                case .center:
-                    .center
-                case .trailing:
-                    .right
-            }
-
-        let resolvedFont = environment.resolvedFont
-
-        // This is definitely what these properties were intended for
-        paragraphStyle.minimumLineHeight = CGFloat(resolvedFont.lineHeight)
-        paragraphStyle.maximumLineHeight = CGFloat(resolvedFont.lineHeight)
-        paragraphStyle.lineSpacing = 0
-
-        return [
-            .foregroundColor: environment.suggestedForegroundColor.resolve(in: environment).nsColor,
-            .font: font(for: resolvedFont),
-            .paragraphStyle: paragraphStyle,
-        ]
-    }
-
-    static func font(for font: Font.Resolved) -> NSFont {
-        let size = CGFloat(font.pointSize)
-        let weight = weight(for: font.weight)
-
-        let nsFont: NSFont
-        switch font.identifier.kind {
-            case .system:
-                switch font.design {
-                    case .default:
-                        nsFont = NSFont.systemFont(ofSize: size, weight: weight)
-                    case .monospaced:
-                        nsFont = NSFont.monospacedSystemFont(ofSize: size, weight: weight)
-                }
-        }
-
-        if font.isItalic {
-            return NSFontManager.shared.convert(nsFont, toHaveTrait: .italicFontMask)
-        } else {
-            return nsFont
-        }
-    }
-
-    private static func weight(for weight: Font.Weight) -> NSFont.Weight {
-        switch weight {
-            case .thin:
-                .thin
-            case .ultraLight:
-                .ultraLight
-            case .light:
-                .light
-            case .regular:
-                .regular
-            case .medium:
-                .medium
-            case .semibold:
-                .semibold
-            case .bold:
-                .bold
-            case .black:
-                .black
-            case .heavy:
-                .heavy
-        }
-    }
-
-    public func createProgressSpinner() -> Widget {
-        let container = NSView()
-        let spinner = NSProgressIndicator()
-        spinner.translatesAutoresizingMaskIntoConstraints = false
-        spinner.isIndeterminate = true
-        spinner.style = .spinning
-        spinner.startAnimation(nil)
-        container.addSubview(spinner)
-        return container
-    }
-
-    public func setSize(
-        ofProgressSpinner widget: Widget,
-        to size: SIMD2<Int>
-    ) {
-        guard Int(widget.frame.size.height) != size.y else { return }
-        setSize(of: widget, to: size)
-        let spinner = NSProgressIndicator()
-        spinner.translatesAutoresizingMaskIntoConstraints = false
-        spinner.isIndeterminate = true
-        spinner.style = .spinning
-        spinner.startAnimation(nil)
-        spinner.widthAnchor.constraint(equalToConstant: CGFloat(size.x)).isActive = true
-        spinner.heightAnchor.constraint(equalToConstant: CGFloat(size.y)).isActive = true
-
-        widget.subviews = []
-        widget.addSubview(spinner)
-    }
-
-    public func createProgressBar() -> Widget {
-        let progressBar = NSProgressIndicator()
-        progressBar.isIndeterminate = false
-        progressBar.style = .bar
-        progressBar.minValue = 0
-        progressBar.maxValue = 1
-        return progressBar
-    }
-
-    public func updateProgressBar(
-        _ widget: Widget,
-        progressFraction: Double?,
-        environment: EnvironmentValues
-    ) {
-        let progressBar = widget as! NSProgressIndicator
-        progressBar.doubleValue = progressFraction ?? 0
-        progressBar.appearance = environment.colorScheme.nsAppearance
-
-        if progressFraction == nil && !progressBar.isIndeterminate {
-            // Start the indeterminate animation
-            progressBar.isIndeterminate = true
-            progressBar.startAnimation(nil)
-        } else if progressFraction != nil && progressBar.isIndeterminate {
-            // Stop the indeterminate animation
-            progressBar.isIndeterminate = false
-            progressBar.stopAnimation(nil)
-        }
-    }
-
-    public func createAlert() -> Alert {
-        NSAlert()
-    }
-
-    public func updateAlert(
-        _ alert: Alert,
-        title: String,
-        actionLabels: [String],
-        environment: EnvironmentValues
-    ) {
-        alert.messageText = title
-        for label in actionLabels {
-            alert.addButton(withTitle: label)
-        }
-    }
-
-    public func showAlert(
-        _ alert: Alert,
-        window: Window?,
-        responseHandler handleResponse: @escaping (Int) -> Void
-    ) {
-        let completionHandler: (NSApplication.ModalResponse) -> Void = { response in
-            guard response != .stop, response != .continue else {
-                return
-            }
-
-            guard response != .abort, response != .cancel else {
-                logger.warning("got abort or cancel modal response, unexpected and unhandled")
-                return
-            }
-
-            let firstButton = NSApplication.ModalResponse.alertFirstButtonReturn.rawValue
-            let action = response.rawValue - firstButton
-            handleResponse(action)
-        }
-
-        if let window {
-            alert.beginSheetModal(
-                for: window,
-                completionHandler: completionHandler
-            )
-        } else {
-            let response = alert.runModal()
-            completionHandler(response)
-        }
-    }
-
-    public func dismissAlert(_ alert: Alert, window: Window?) {
-        if let window {
-            window.endSheet(alert.window)
-        } else {
-            NSApplication.shared.stopModal()
-        }
-    }
-
-    public func showOpenDialog(
-        fileDialogOptions: FileDialogOptions,
-        openDialogOptions: OpenDialogOptions,
-        window: Window?,
-        resultHandler handleResult: @escaping (DialogResult<[URL]>) -> Void
-    ) {
-        let panel = NSOpenPanel()
-        panel.message = fileDialogOptions.title
-        panel.prompt = fileDialogOptions.defaultButtonLabel
-        panel.directoryURL = fileDialogOptions.initialDirectory
-        panel.showsHiddenFiles = fileDialogOptions.showHiddenFiles
-        panel.allowsOtherFileTypes = fileDialogOptions.allowOtherContentTypes
-
-        // TODO: allowedContentTypes
-
-        panel.allowsMultipleSelection = openDialogOptions.allowMultipleSelections
-        panel.canChooseFiles = openDialogOptions.allowSelectingFiles
-        panel.canChooseDirectories = openDialogOptions.allowSelectingDirectories
-
-        let handleResponse: (NSApplication.ModalResponse) -> Void = { response in
-            guard response != .continue else {
-                return
-            }
-
-            if response == .OK {
-                handleResult(.success(panel.urls))
-            } else {
-                handleResult(.cancelled)
-            }
-        }
-
-        if let window {
-            panel.beginSheetModal(for: window, completionHandler: handleResponse)
-        } else {
-            let response = panel.runModal()
-            handleResponse(response)
-        }
-    }
-
-    public func showSaveDialog(
-        fileDialogOptions: FileDialogOptions,
-        saveDialogOptions: SaveDialogOptions,
-        window: Window?,
-        resultHandler handleResult: @escaping (DialogResult<URL>) -> Void
-    ) {
-        let panel = NSSavePanel()
-        panel.message = fileDialogOptions.title
-        panel.prompt = fileDialogOptions.defaultButtonLabel
-        panel.directoryURL = fileDialogOptions.initialDirectory
-        panel.showsHiddenFiles = fileDialogOptions.showHiddenFiles
-        panel.allowsOtherFileTypes = fileDialogOptions.allowOtherContentTypes
-
-        // TODO: allowedContentTypes
-
-        panel.nameFieldLabel = saveDialogOptions.nameFieldLabel ?? panel.nameFieldLabel
-        panel.nameFieldStringValue = saveDialogOptions.defaultFileName ?? ""
-
-        let handleResponse: (NSApplication.ModalResponse) -> Void = { response in
-            guard response != .continue else {
-                return
-            }
-
-            if response == .OK {
-                handleResult(.success(panel.url!))
-            } else {
-                handleResult(.cancelled)
-            }
-        }
-
-        if let window {
-            panel.beginSheetModal(for: window, completionHandler: handleResponse)
-        } else {
-            let response = panel.runModal()
-            handleResponse(response)
-        }
-    }
-
-    public func createDatePicker() -> NSView {
-        let datePicker = CustomDatePicker()
-        datePicker.delegate = datePicker.strongDelegate
-        return datePicker
-    }
-
-    // Depending on the calendar, era is either necessary or must be omitted. Making the wrong
-    // choice for the current calendar means the cursor position is reset after every keystroke. I
-    // know of no simple way to tell whether NSDatePicker requires or forbids eras for a given
-    // calendar, so in lieu of that I have hardcoded the calendar identifiers.
-    private let calendarsRequiringEra: Set<Calendar.Identifier> = [
-        .buddhist,
-        .coptic,
-        .ethiopicAmeteAlem,
-        .ethiopicAmeteMihret,
-        .indian,
-        .islamic,
-        .islamicCivil,
-        .islamicTabular,
-        .islamicUmmAlQura,
-        .japanese,
-        .persian,
-        .republicOfChina,
-    ]
-
-    public func updateDatePicker(
-        _ datePicker: NSView,
-        environment: EnvironmentValues,
-        date: Date,
-        range: ClosedRange<Date>,
-        components: DatePickerComponents,
-        onChange: @escaping (Date) -> Void
-    ) {
-        let datePicker = datePicker as! CustomDatePicker
-
-        datePicker.isEnabled = environment.isEnabled
-        datePicker.textColor = environment.suggestedForegroundColor.resolve(in: environment).nsColor
-
-        // If the time zone is set to autoupdatingCurrent, then the cursor position is reset after
-        // every keystroke. Thanks Apple
-        datePicker.timeZone =
-            environment.timeZone == .autoupdatingCurrent ? .current : environment.timeZone
-
-        // A couple properties cause infinite update loops if we assign to them on every update, so
-        // check their values first.
-        if datePicker.calendar != environment.calendar {
-            datePicker.calendar = environment.calendar
-        }
-
-        if datePicker.dateValue != date {
-            datePicker.dateValue = date
-        }
-
-        var elementFlags: NSDatePicker.ElementFlags = []
-        if components.contains(.date) {
-            elementFlags.insert(.yearMonthDay)
-            if calendarsRequiringEra.contains(environment.calendar.identifier) {
-                elementFlags.insert(.era)
-            }
-        }
-        if components.contains(.hourMinuteAndSecond) {
-            elementFlags.insert(.hourMinuteSecond)
-        } else if components.contains(.hourAndMinute) {
-            elementFlags.insert(.hourMinute)
-        }
-
-        if datePicker.datePickerElements != elementFlags {
-            datePicker.datePickerElements = elementFlags
-        }
-
-        datePicker.strongDelegate.onChange = onChange
-
-        datePicker.minDate = range.lowerBound
-        datePicker.maxDate = range.upperBound
-
-        datePicker.datePickerStyle =
-            switch environment.datePickerStyle {
-                case .automatic, .compact:
-                    .textFieldAndStepper
-                case .graphical:
-                    .clockAndCalendar
-            }
-    }
-}
-
-final class NSCustomMenuItem: NSMenuItem {
-    /// This property's only purpose is to keep a strong reference to the wrapped
-    /// action so that it sticks around for long enough to be useful.
-    var actionWrapper: Action?
-
-    var isOn: Bool {
-        get { state == .on }
-        set { state = newValue ? .on : .off }
-    }
 }
 
 // TODO: Update all controls to use this style of action passing, seems way nicer
@@ -1457,63 +482,6 @@ final class Action: NSObject {
 
     @objc func run() {
         action()
-    }
-}
-
-class NSCustomTableView: NSTableView {
-    var customDelegate = NSCustomTableViewDelegate()
-}
-
-class NSCustomTableViewDelegate: NSObject, NSTableViewDelegate, NSTableViewDataSource {
-    var widgets: [AppKitBackend.Widget] = []
-    var rowHeights: [Int] = []
-    var columnIndices: [ObjectIdentifier: Int] = [:]
-    var rowCount = 0
-    var columnCount = 0
-    var allowSelections = false
-    var selectionHandler: ((Int) -> Void)?
-
-    func numberOfRows(in tableView: NSTableView) -> Int {
-        return rowCount
-    }
-
-    func tableView(_ tableView: NSTableView, heightOfRow row: Int) -> CGFloat {
-        return CGFloat(rowHeights[row])
-    }
-
-    func tableView(
-        _ tableView: NSTableView,
-        viewFor tableColumn: NSTableColumn?,
-        row: Int
-    ) -> NSView? {
-        guard let tableColumn else {
-            logger.warning("no column provided")
-            return nil
-        }
-        guard let columnIndex = columnIndices[ObjectIdentifier(tableColumn)] else {
-            logger.warning("NSTableView asked for value of non-existent column")
-            return nil
-        }
-        return widgets[row * columnCount + columnIndex]
-    }
-
-    func tableView(
-        _ tableView: NSTableView,
-        selectionIndexesForProposedSelection proposedSelectionIndexes: IndexSet
-    ) -> IndexSet {
-        if allowSelections {
-            selectionHandler?(proposedSelectionIndexes.first!)
-            return proposedSelectionIndexes
-        } else {
-            return []
-        }
-    }
-
-    func tableView(_ tableView: NSTableView, rowViewForRow row: Int) -> NSTableRowView? {
-        let view = NSTableRowView()
-        view.wantsLayer = true
-        view.layer?.cornerRadius = 5
-        return view
     }
 }
 
@@ -1574,120 +542,6 @@ final class ObjectAssociation<T: Any> {
                 newValue,
                 policy.rawValue
             )
-        }
-    }
-}
-
-// Source: https://gist.github.com/sindresorhus/3580ce9426fff8fafb1677341fca4815
-extension NSControl {
-    typealias ActionClosure = ((NSControl) -> Void)
-    typealias EditClosure = ((NSTextField) -> Void)
-
-    @MainActor
-    struct AssociatedKeys {
-        static let onActionClosure = ObjectAssociation<ActionClosure>()
-        static let onEditClosure = ObjectAssociation<EditClosure>()
-    }
-
-    @objc
-    func callClosure(_ sender: NSControl) {
-        onAction?(sender)
-    }
-
-    var onAction: ActionClosure? {
-        get {
-            return AssociatedKeys.onActionClosure[self]
-        }
-        set {
-            AssociatedKeys.onActionClosure[self] = newValue
-            action = #selector(callClosure)
-            target = self
-        }
-    }
-}
-
-class NSCustomSplitView: NSSplitView {
-    var resizingDelegate: NSSplitViewResizingDelegate?
-}
-
-class NSSplitViewResizingDelegate: NSObject, NSSplitViewDelegate {
-    var resizeHandler: (() -> Void)?
-    var leadingWidth = 0
-    var minimumLeadingWidth = 0
-    var maximumLeadingWidth = 0
-    var isFirstUpdate = true
-    /// Tracks whether AppKit is resizing the side bar (as opposed to the user resizing it).
-    var appKitIsResizing = false
-
-    func setResizeHandler(_ handler: @escaping () -> Void) {
-        resizeHandler = handler
-    }
-
-    func splitView(_ splitView: NSSplitView, shouldAdjustSizeOfSubview view: NSView) -> Bool {
-        appKitIsResizing = true
-        return true
-    }
-
-    func splitViewDidResizeSubviews(_ notification: Notification) {
-        appKitIsResizing = false
-        let splitView = notification.object! as! NSSplitView
-        let paneWidths = splitView.subviews.map(\.frame.width).map { width in
-            Int(width.rounded())
-        }
-        let previousWidth = leadingWidth
-        leadingWidth = paneWidths[0]
-
-        // Only call the handler if the side bar has actually changed size.
-        if leadingWidth != previousWidth {
-            resizeHandler?()
-        }
-    }
-
-    func splitView(
-        _ splitView: NSSplitView,
-        constrainMinCoordinate proposedMinimumPosition: CGFloat,
-        ofSubviewAt dividerIndex: Int
-    ) -> CGFloat {
-        if dividerIndex == 0 {
-            return CGFloat(minimumLeadingWidth)
-        } else {
-            return proposedMinimumPosition
-        }
-    }
-
-    func splitView(
-        _ splitView: NSSplitView,
-        constrainMaxCoordinate proposedMaximumPosition: CGFloat,
-        ofSubviewAt dividerIndex: Int
-    ) -> CGFloat {
-        if dividerIndex == 0 {
-            return CGFloat(maximumLeadingWidth)
-        } else {
-            return proposedMaximumPosition
-        }
-    }
-
-    func splitView(_ splitView: NSSplitView, resizeSubviewsWithOldSize oldSize: NSSize) {
-        splitView.adjustSubviews()
-
-        if isFirstUpdate {
-            splitView.setPosition(max(200, CGFloat(minimumLeadingWidth)), ofDividerAt: 0)
-            isFirstUpdate = false
-        } else {
-            let newWidth = splitView.subviews[0].frame.width
-            // If AppKit is trying to automatically resize our side bar (e.g. because the split
-            // view has changed size), only let it do so if not doing so would put out side bar
-            // outside of the allowed bounds.
-            if appKitIsResizing
-                && leadingWidth >= minimumLeadingWidth
-                && leadingWidth <= maximumLeadingWidth
-            {
-                splitView.setPosition(CGFloat(leadingWidth), ofDividerAt: 0)
-            } else {
-                // Magic! Thanks https://stackoverflow.com/a/30494691. This one line fixed all
-                // of the split view resizing issues.
-                splitView.setPosition(newWidth, ofDividerAt: 0)
-            }
         }
     }
 }
@@ -1774,100 +628,5 @@ final class NSCustomApplicationDelegate: NSObject, NSApplicationDelegate {
 
     func application(_ application: NSApplication, open urls: [URL]) {
         onOpenURLs?(urls)
-    }
-}
-
-/// A scroll view with scrolling gestures disabled. Used as a dummy scroll view to
-/// allow us to properly set the width of NSTableView (had some weird issues).
-final class NSDisabledScrollView: NSScrollView {
-    override func scrollWheel(with event: NSEvent) {
-        self.nextResponder?.scrollWheel(with: event)
-    }
-}
-
-final class CustomDatePicker: NSDatePicker {
-    var strongDelegate = CustomDatePickerDelegate()
-}
-
-final class CustomDatePickerDelegate: NSObject, NSDatePickerCellDelegate {
-    var onChange: ((Date) -> Void)?
-
-    func datePickerCell(
-        _: NSDatePickerCell,
-        validateProposedDateValue proposedDateValue: AutoreleasingUnsafeMutablePointer<NSDate>,
-        timeInterval _: UnsafeMutablePointer<TimeInterval>?
-    ) {
-        onChange?(proposedDateValue.pointee as Date)
-    }
-}
-
-final class RadioGroup: NSStackView {
-    private var buttons: [NSButton]
-    var onChange: ((Int?) -> Void)?
-
-    override var intrinsicContentSize: NSSize {
-        buttons.reduce(
-            into: NSSize(width: 0.0, height: max(0.0, spacing * Double(buttons.count - 1)))
-        ) { partialResult, button in
-            let buttonIntrinsicSize = button.intrinsicContentSize
-            partialResult.width = max(partialResult.width, buttonIntrinsicSize.width)
-            partialResult.height += buttonIntrinsicSize.height
-        }
-    }
-
-    init() {
-        self.buttons = []
-        super.init(frame: .zero)
-        self.orientation = .vertical
-        self.alignment = .leading
-        self.setAccessibilityRole(.radioGroup)
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("not used")
-    }
-
-    func update(options: [String], environment: EnvironmentValues) {
-        for i in 0..<min(buttons.count, options.count) {
-            buttons[i].attributedTitle = AppKitBackend.attributedString(
-                for: options[i],
-                in: environment
-            )
-            buttons[i].isEnabled = environment.isEnabled
-        }
-
-        if options.count > buttons.count {
-            for i in buttons.count..<options.count {
-                let button = NSButton()
-                button.attributedTitle = AppKitBackend.attributedString(
-                    for: options[i],
-                    in: environment
-                )
-                button.isEnabled = environment.isEnabled
-                button.target = self
-                button.action = #selector(buttonClicked(sender:))
-                button.tag = i
-                button.setButtonType(.radio)
-                addArrangedSubview(button)
-                buttons.append(button)
-            }
-        } else {
-            for i in (options.count..<buttons.count).reversed() {
-                removeView(buttons[i])
-                buttons.remove(at: i)
-            }
-        }
-    }
-
-    func setSelectedIndex(to index: Int?) {
-        if let index {
-            buttons[index].state = .on
-        } else {
-            buttons.forEach { $0.state = .off }
-        }
-    }
-
-    @objc func buttonClicked(sender: NSButton) {
-        onChange?(sender.tag)
     }
 }
